Hard code rather than use facts because Debian makes me cry
I wanted to make a fact for the location in which python packages should
be installed, but because the way Debian does it by patching
easy_install/pip and not actually setting it in the python --configure
options, you can't determine that information until after pip is
installed.  This doesn't work since facts are loaded prior to the puppet
run.

Once again, thanks Debian for thinking you know better than the people
who write the software.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>